Fixed incomplete type compilation errors

+++ nestkernel/source_table_position.cpp
@@ -29,19 +29,11 @@ nest::SourceTablePosition::decrease()
       syn_id = kernel().model_manager.get_num_connection_models() - 1;
     }
 
-    local_target_connection_id = kernel()
-                                   .node_manager.get_local_nodes( tid )
-                                   .get_node_by_index( local_target_node_id )
-                                   ->get_num_conn_type_sources( syn_id )
-      - 1;
+    local_target_connection_id = kernel().node_manager.get_local_nodes( tid ).get_node_by_index( local_target_node_id )->get_num_conn_type_sources( syn_id ) - 1;
   }
 
   // if the found index is not valid, decrease further
-  if ( local_target_connection_id
-    == kernel()
-         .node_manager.get_local_nodes( tid )
-         .get_node_by_index( local_target_node_id )
-         ->get_num_conn_type_sources( syn_id ) )
+  if ( local_target_connection_id == static_cast<long>(kernel().node_manager.get_local_nodes( tid ).get_node_by_index( local_target_node_id )->get_num_conn_type_sources( syn_id ) ) )
   {
     decrease();
   }
@@ -52,23 +44,19 @@ nest::SourceTablePosition::increase()
 {
   // first try finding a valid index by only increasing target-local connection id
   ++local_target_connection_id;
-  if ( local_target_connection_id
-    == kernel()
-         .node_manager.get_local_nodes( tid )
-         .get_node_by_index( local_target_node_id )
-         ->get_num_conn_type_sources( syn_id ) )
+  if ( local_target_connection_id == static_cast<long>(kernel().node_manager.get_local_nodes( tid ).get_node_by_index( local_target_node_id )->get_num_conn_type_sources( syn_id ) ) )
   {
     // then try finding a valid index by increasing synapse index
     ++syn_id;
-    if ( syn_id == kernel().model_manager.get_num_connection_models() )
+    if ( syn_id == static_cast<long>(kernel().model_manager.get_num_connection_models() ) )
     {
       // then try finding a valid index by increasing target node id
       ++local_target_node_id;
-      if ( local_target_node_id == kernel().node_manager.get_local_nodes( tid ).size() )
+      if ( local_target_node_id == static_cast<long>(kernel().node_manager.get_local_nodes( tid ).size() ) )
       {
         // then try finding a valid index by increasing thread index
         ++tid;
-        if ( tid == kernel().vp_manager.get_num_threads() )
+        if ( tid == static_cast<long>(kernel().vp_manager.get_num_threads() ) )
         {
           return; // reached the end without finding a valid entry
         }
@@ -83,11 +71,7 @@ nest::SourceTablePosition::increase()
   }
 
   // if the found index is still not valid, increase further
-  if ( local_target_connection_id
-    == kernel()
-         .node_manager.get_local_nodes( tid )
-         .get_node_by_index( local_target_node_id )
-         ->get_num_conn_type_sources( syn_id ) )
+  if ( local_target_connection_id == static_cast<long>(kernel().node_manager.get_local_nodes( tid ).get_node_by_index( local_target_node_id )->get_num_conn_type_sources( syn_id ) ) )
   {
     increase();
   }
